+++ lib/Pages/equip_page.dart
