+++ lib/splash.dart
@@ -1,4 +1,4 @@
-import 'package:dragcon/Pages/LoginScreen.dart';
+import 'package:dragcon/Pages/login_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_animation_progress_bar/flutter_animation_progress_bar.dart';
