+++ lib/web_api/connection/task_connection.dart
@@ -11,7 +11,7 @@ class TaskConnection {
   final apiService = ApiService();
 
   Future<TaskDto> getTaskById(int id) async {
-    final Response response = await apiService.makeApiGetRequest('$apiHost/api/tasks/$id');
+    final Response response = await apiService.makeApiGetRequest('$apiHost/tasks/$id');
 
     if (response.statusCode == 404) {
       throw CantFetchDataException();
@@ -22,16 +22,17 @@ class TaskConnection {
 
   patchTaskById(int id, TaskDto taskDto) async {
     final Response response = await apiService.patch('$apiHost/api/tasks/$id', taskDto);
+    print(response.body);
     return response.statusCode;
   }
 
   addNewTask(TaskDto taskDto) async {
-    final Response response = await apiService.post('$apiHost/api/tasks', taskDto);
+    final Response response = await apiService.post('$apiHost/tasks', taskDto);
     return response.statusCode;
   }
 
   deleteTask(int id) async {
-    final Response response = await apiService.delete('$apiHost/api/tasks/$id');
+    final Response response = await apiService.delete('$apiHost/tasks/$id');
     return response.statusCode;
   }
 
@@ -40,8 +41,9 @@ class TaskConnection {
     int pageNumber = 1;
 
     while (true) {
+      print("XD");
       var items = await _getDtosByPageNumberAndId(pageNumber, id);
-
+      print(items);
       if (items.isEmpty) {
         break;
       } else {
@@ -49,6 +51,7 @@ class TaskConnection {
       }
 
       dtos.addAll(items);
+      return dtos;
     }
     return dtos;
   }
@@ -58,13 +61,13 @@ class TaskConnection {
     int id,
   ) async {
     var response = await apiService.makeApiGetRequest(
-      '$apiHost/api/tasks-by-ekipa/$id?page=$pageNumber',
+      '$apiHost/tasks-by-ekipa/$id?page=$pageNumber',
     );
 
-    if (response.statusCode == 201) {
+    if (response.statusCode != 400) {
       var decodedBody = json.decode(response.body);
-      var items = ResponseHelper.items(decodedBody);
-      return items.map((e) => TaskDto.fromJson(e)).toList();
+      var taskList = List<Map<String, dynamic>>.from(decodedBody);
+      return taskList.map((e) => TaskDto.fromJson(e)).toList();
     } else {
       throw CantFetchDataException();
     }
