+++ lib/web_api/service/api_service.dart
@@ -7,6 +7,7 @@ class ApiService {
   // final SharedPreference sharedPreference = GetIt.I<SharedPreference>();
 
   Future<http.Response> makeApiGetRequest(String uri) async {
+    print(uri);
     return http.get(
       Uri.parse(uri),
       headers: <String, String>{
@@ -20,11 +21,21 @@ class ApiService {
     DtoToJsonInterface toJsonDto,
   ) async {
     // var token = await sharedPreference.getApiToken();
-
+    print(uri);
+    print(http.patch(
+      Uri.parse(uri),
+      headers: <String, String>{
+        'Content-Type': 'application/merge-patch+json; charset=UTF-8',
+        'Accept': 'application/ld+json', // Dodaj nagłówek Accept
+        // 'Authorization': 'Bearer $token'
+      },
+      body: jsonEncode(toJsonDto.toJson()),
+    ));
     return http.patch(
       Uri.parse(uri),
       headers: <String, String>{
         'Content-Type': 'application/merge-patch+json; charset=UTF-8',
+        'Accept': 'application/ld+json', // Dodaj nagłówek Accept
         // 'Authorization': 'Bearer $token'
       },
       body: jsonEncode(toJsonDto.toJson()),
@@ -71,5 +82,4 @@ class ApiService {
       },
     );
   }
-
 }
